Handle conflicting experiments when opening
When opening a trace with the same name as another experiment, the experiment
that contains it will have a conflict on the server. The client should
retry posting it with a different name until it succeeds or another
error occurs.

The title of the trace viewer should also be updated to represent the
actual name of the experiment.

Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>