Added Gitpod config
The config consists of two files. They were copied and slightly
modified from the upstream Eclipse Theia project:

.gitpod.yml and .gitpod.dockerfile

See LICENSE.gitpod.md for details about the license of these files.

As well I let Gitpod add it's README banner, so that a visitor can
try the repo's example application at the click of the button.
The only way this would be better is to have a trace to open.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>