Add npm script to download trace server and sample traces
The trace server download will work when the update
in the Trace Compass CI matches.

Gitpod config switched to produce the electron version
of the example trace viewer application. The browser
version does not currently work when the frontend and
backend of the app are on different machines.

To access the example application on Gitpod, wait for
it to build, and when done open port 6080 from the
right side of the status bar. This will open a new
browser tab or window that will access the vnc server
that runs in your container. This way of doing things
introduces significant lag vs the broser version. That's
unavoidable. Even if not ideal, one can test a PR in 3
mins instead of 15 this way.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>