update repo's configs for Electron 9 / node 12
Fixes #106

A few little tweaks are necessary to accommodate the new Electron and node versions. ATM the Electron example app fails at runtime when running on Gitpod, because of a missing a dynamic library:

/workspace/theia-trace-extension/node_modules/electron/dist/electron: error while loading shared libraries: libgbm.so.1: cannot

We need to update the Gitpod config so the dockerfile used for this repoi will contain a new Ubuntu package, required for Electron 9. This config file originally came from the main Theia repo, and the updated version is the same. See LICENSE.gitpod.md for more details.

Security note: in order for the Electron version of the example application to work in Gitpod (probably Docker in general), two command-line arguments are required. These are not to be used in production since they decrease security. However they should only be needed momentarily for tests in Gitpod, until we can use the browser version in there and in any other future Cloud-based applications that will use the "trace" extension:

  --no-sandbox
  --ignore-gpu-blacklist

These two arguments are used only in Gitpod, from `.gitpod.yml`

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>