Migrating the repo to a vscode extension
To build: first from the root run `yarn`. It will build all the packages
for the vscode extension to work.

After modifications to the code, the corresponding workspace needs to be
built again. In packages/vscode-extension are 2 build commands that can
also be watch: build/watch:extension to build the core vscode extension
and build/watch:react to build the react app for the trace viewer.

To run the extension from vscode itself, press F5 in vscode, or
selecting the launch configuration called `Launch Extension`. The
configuration does not automatically build anything.

Signed-off-by: Quoc-Hao TRAN <quochao.tran.up@gmail.com>
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>