Split viewer-prototype in packages
In prevision of the development of a vscode extension, the code in the
viewer-prototype has been mostly moved to 2 new packages:

* @trace-viewer/base: Queries the trace server for traces, experiments
and possible outputs
* @trace-viewer/react-components: Contains the various types of
visualization for trace data, queries the trace server for output data
to display

Those 2 packages can eventually be published to npm.

In the viewer-prototype directory only remains the theia specific module
contributions. So the application still works as before using theia,
it's just that most of the classes used by the trace-viewer have been
moved to new npm packages.

Signed-off-by: Geneviève Bastien <gbastien@versatic.net>