Support inband OutputElementStyle for annotations
To provide annotation markers the output styles needed to defined
in the list of style provided through the fetch style method which
are referenced by the parent style key in the annotation itself.

This patch makes it possible to supply the OutputElementStyle inside
the annotation without using the parent style key.

Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>