fix and refactor plugins
Share all plugins between both the browser and the electron example
applications. The plugin list is maintained under the root
`package.json` and they are installed in `examples/plugins`.

`cdt-gdb-vscode` was unpublished from openvsx so it had to be removed.

`docker`, `html` and `markdown` were bugged so they were removed.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>