upgrade dependencies
The previous locked Theia version was 1.6.0, which is now missing quite
a lot of new VS Code APIs. This doesn't play well with the previous
commit that updates the plugin URLs to newer versions, as the plugins
would try to use those new APIs.

This commit upgrades all dependencies without changing our ranges.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>