Show view message when there is no data to provide
Some views weren't able to display any data after the completion of the trace's analysis.
The problem was due to the fact that some provided models are empty server side because
they were no relevents trace events to display.

We choose to display a message indicating that there are no data to display in those cases
and update the state of the analysis client side. It will allow a better clarity of the user
experience.

Fixes #406
Signed-off-by: Jeff Godonou <jeffalecgodonou@gmail.com>