cleaned-out license-check-baseline.json
This file acts as a filter, that contains dependencies that
fail the liocense check but are "known". This permits to ignore
these, and be warned when any new dependency that fails the check
is added, e.g. in a PR.

For now I have kept upstream dependencies that are relevant here.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>