Clarify package info cleanup not related to keytar
Since the convention is to put the comments that clarify the
installation commands above rather than next to the command, it looks
like the comment labels a section of possibly more than one command.
When the install and remove commands aren't separated by anything it
looks like both commands are part of setting up the keytar package.

For comparison, this is what it looked like before:

    # keytar
    && sudo apt-get install -y libsecret-1-dev \
    && sudo rm -rf /var/lib/apt/lists/*

Add comment between commands to clarify that they're unrelated.

Signed-off-by: Erica Bugden <erica.bugden@gmail.com>