Let timeline-chart create gap states
Do not add gap states to the row model, but include all states including
null states (states without a style). Add the optional gap style for a
row to the row model. This will allow the timeline-chart to create its
own gaps dynamically as needed for its current view range.

Signed-off-by: Patrick Tasse <patrick.tasse@ericsson.com>