Move test instructions from the packages README.md to the root README.md
The package READEME.md is used on npm registry to give users information
about the library. Instructions for developers that develop the
theia-trace-extensions need to go in the root README.md

Update test script in the root package.json to allow forwarding
parameters to the packages test script execution. With this the test
can be executed from the root directory.

Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>