Don't execute open command if TraceViewerWidget is already open
This avoids duplicate TraceViewerWidget tabs when 1) open a trace from
the file explorer and 2) switch between traces in the open traces
widget in the trace explorer. For 1) the widget is created with an URI
(file path) and for 2) the widget is created with an existing trace
UUID. Internally in Theia these causes 2 widgets to be opened because
it uses the URI and the TraceViewerWidgetOptions (incl. UUID) to create
a key for an internal map. For both cases the key is different.

Avoiding executing the open command makes sure that the existing widget
is used, no matter which way it was opened.

Fixes #450

Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>