Scatter chart
Data sets with scattered points, like the "latency vs. time" view, are
now plotted in a scatter chart. The type of data set is defined by the
string "scatter" in the output id.

The scatter chart is to remain not available to the user until the
tooltip is implemented.

This pull request replaces
https://github.com/theia-ide/theia-trace-extension/pull/399.

Contributes towards fixing #45.

Signed-off-by: Rodrigo Pinto <rodrigo.pinto@calian.ca>