Add undo/redo (hardcode)
Currently, the timeRange is stored in undo/redo stack so that the user is able to redo/undo zoom in/zoom out/pan left/pan right. I tried to make it generic by storing the entire context instead of just the timeRange so that other setting such as backgroundTheme can be undone/redone as well. I tried undoStack.push(this) but it doesn’t seem to be the right way.

Any advice is appreciated!
Signed-off-by: Yining Wang <yining.wang@ericsson.com>