Set keyboard focus on graph/table when selecting an analysis
fixes #414

When clicking on an analysis in the Available Analysis list, the analysis
is scrolled into view and given keyboard focus. When clicking on an analysis
title in the trace tab, keyboard focus should be the analysis' graph/chart region.
If a View that is already open is selected in the sidebar, the View's title area
should pulse to temporarily bring attention to the View panel. 

Signed-off-by: hriday-panchasara <hriday.panchasara@ericsson.com>