Fix link between table event and XY chart
Opening an Events Table and a XY chart like "CPU Usage" and clicking on
a row in the table was supposed to show a vertical line in the XY chart,
but this was not happening.

Selecting several rows in the table also was supposed to show a blue
selection area in the chart, which was not working either.

The selection would only work if, apart from opening an Events Table and
a XY chart, a status chart was opened, like the "Resources Chart".

Now all selections work.

Fixes #448.

Signed-off-by: Rodrigo Pinto <rodrigo.pinto@calian.ca>