Add undo/redo functionality(bug free) Since it is not allowed to assign 'this', the value of viewRange cannot be changed in the function. Thus, I change the return type of undo/redo function from void to bigint[]. In specific, it returns a list of bigint which are start and end of viewRange. Then, this.unitController.viewRange can be changed after the function is called.
I tried a few ways to expand it to a generic undo/redo function. However, none of it works.

1. To store TraceContextComponent
I tried to store 'this' which is TraceContextComponent in the undoStack. However, it is similar to passing by reference, so no value can be changed if we do 'this = undoStack.pop()'.

2. To store TraceContextProps
I also used undoStack to store TraceContextProps, but 'this.props = undoStack.pop()' is not allowed because props is read only.

3. To store state
I used 'this.setState(currState: prevState)' but this.state does not include every property we want to keep track of. For example, this.state does not include unitController which contains viewRange.

I am sorry about the late reply. Since I do not have the access to the previous email, I will use the current account in the future. Also, happy lunar new year!!!!

Signed-off-by: Yining Wang <yining.wang@ericsson.com>