GraphQL ADR: Prototype resolvers for tsp-client.ts
Base this on below [1]'s instructions and resulting skeleton files,
amended for TSP's /health and (limited) /traces endpoints solely.

Include an executable bash ./test script with its assumptions documented
in it. Join an optional VS Code launch configuration that can replace
running [2] (below) manually; use it for occasional debugging too.

Using these hereby added resolvers depends on a running default
trace-server locally. Per [1], resolvers are in JavaScript as they only
demonstrate this limited scope. TypeScript may be used if ever expanding
such use of resolvers, depending on if GraphQL ever gets in TSP or not.

Refer to [3] from this repository, for how to call client's checkHealth;
refer to [4] for fetchTraces but only return their amount. More on how
to implement resolvers, including more advanced ones, is in [5]. REST
endpoints can otherwise be resolved through a data source [6], but that
approach remains out of this small prototype scope for now.

Document the nature of this prototype, along with steps, in the ADR. Add
the related reference about resolving Graph queries with REST endpoints.

[1] https://www.apollographql.com/docs/apollo-server/getting-started/
[2] node index.js
[3] trace-viewer-contribution.ts
[4] trace-manager.ts
[5] https://www.apollographql.com/docs/apollo-server/data/resolvers/
[6] https://www.apollographql.com/docs/apollo-server/data/data-sources/#restdatasource-reference

Signed-off-by: Marco Miller <marco.miller@ericsson.com>