Launch: Add an all-tests configuration for Jest
Add a launch configuration that runs all the Jest tests, without having
to open (or be on) any one of the __tests__ files. Make the new launcher
rely on the existing 'yarn test' cross-repository (standard) command.

Except for the non-applicable or replaced options, use the same options
as the ones from the sibling Jest Current File configuration. Similar to
the latter that allows debugging from a specific test suite, this new
configuration enables debugging from every (any) test suite, without
having to know which one(s) to select up-front.

To reproduce passing tests with this new launcher,
- in VS Code, go to the Run and Debug panel; then,
- select the Jest All Tests launch configuration from that menu;
- either Start Debugging or Run Without Debugging.

Signed-off-by: Marco Miller <marco.miller@ericsson.com>