Fix invalid path on Windows when opening traces
Currently, when opening traces on Windows, Theia returns the file path
starting with a /. For example, instead of C:/, Theia returns /C://.
The path is an invalid Windows path, and prevent the trace extension
from opening traces. This is a known issue that requires a work around.

This commit resolves this issue by using FileUri to convert the original
file path returned by Theia to a platform specific path. This work
around will temporarily allow traces to be opened in Windows environment.

Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>