Add ADR for Hotjar investigation
The trace extension would like to obtain user feedback to
improve its UI/UX. This commit adds an ADR about using Hotjar
investigation to track user happiness for the Theia trace
extension. It is a thorough analysis of Hotjar, its limitations
and the constraints of the trace extension. The conclusion is
that we cannot use Hotjar because the extension cannot have any
commercial dependencies.

Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>