Update open trace icon to avoid confusion
Currently, the trace extension uses the 'Open folder'
icon for the open trace button. This can be confusing
for users, since they can view the action as closing the
current trace and opening a new one. The expected
behavior of this action is opening a new trace along
side the old ones. This commit updates the icon of the
open trace button to 'Folder plus' to avoid this confusion.
Since the 'Folder plus' icon is not yet supported by the
current font-awesome version of Theia (4.7.0), this commit
opts for the same icon provided by @vscode/codicons instead.

Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>