D3 ADR: move code example from proto file into ADR
The `0003-D3-charts` ADR used to refer to the file
`0003/proto_xy-output-component.tsx`, which contained an implementation
of D3 charts for the `xy-output-component.tsx` file.

This proto file would appear into searches while navigating in the
project, adding duplicated results that were not usefull for developers.

To solve this issue, this PR imports the code from the proto file into
the ADR, which stops its contents from appearing into searches for ts
and tsx files.

Signed-off-by: Rodrigo Pinto <rodrigo.pinto@calian.ca>