Select chart components from a selection in the events table
Select corresponding chart components from the events table.

Clicking on a line in the events table will cause a chart like "Thread
Status" to scroll vertically to the row which has the element that
corresponds to the clicked event.

For each line in the timegraph tree, parse the metadata in the
TimeGraphEntry received from the server and try to find matches with the
key / values pairs of the clicked event. It will select the entry which
has the most matches.

When extending the selection, only the last selected event will be
used to broadcast the selection (multi-line selection).

This PR requires the following pull requests:
https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/193246
https://github.com/eclipse-cdt-cloud/tsp-typescript-client/pull/63
https://github.com/eclipse-cdt-cloud/timeline-chart/pull/204

The corresponding Trace Server Protocol (TSP) update:
https://github.com/eclipse-cdt-cloud/trace-server-protocol/pull/83

Signed-off-by: Rodrigo Pinto <rodrigo.pinto@calian.ca>
Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>