Select chart components from the events table
Select corresponding chart components from the events table.

Clicking on a line in the events table will cause a chart like "Thread
Status" to scroll vertically to the row which has the element that
corresponds to the clicked event.

Only works if the events contains matching information about start time,
PID, TID or unique process name.

This PR requires the following pull requests:

https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/193246
https://github.com/eclipse-cdt-cloud/tsp-typescript-client/pull/62

Signed-off-by: Rodrigo Pinto <rodrigo.pinto@calian.ca>