WIP: propose a framework for documentation
A common problem for SW projects is how to create
documentation (both user oriented, describing how
to use the tool, and developer oriented, describing
concepts such as “design specifications”, “architecture
descriptions”, "adrs").

This commit aims to show how we could create developer
oriented documentation. It add a framework that picks the
existing adrs markdown files, and deploy them to a static
webpage.

Benefits of the proposed approach:

- documentation is in the same repo as the code (no
  risk of spreading it in different tools/repos);

- documentation is versioned in the same way as code;

- documentation can be still be read/visualized on github,
  but it can also be deployed locally on the machine
  where the repo has been cloned, or deployed to github
  page or similar;

- the markdown files representing the documentation
  (in this case the adr) are not modified;

- this approach could be common between multiple
  projects under the eclipse-cdt-cloud group;

The same approach could be used to generate user oriented
documentation, but it is outside the scope of this commit.

See the README file to get going.

Signed-off-by: Francesco Robino <francesco.robino@ericsson.com>