New foldable sidebar one row per folder
Before this commit, the sidebar was generated with
one row per md file.

After this commit, the sidebar is generated with
one row per folder containing md files.

This is not the final solution, but just shows how
we can play with sidebar generation depending on
how we want to distribute docs in the project
or if we want to use front matter.