Display data as it is made available
Allows the overview chart to display data as it is being loaded.

The effect is achieved by replacing `debounce` with a timeout function
that calls `updateXY()` once every 3 seconds. Shorter intervals cause
`fetchXY()` to bee called too many times in large traces, and generates
errors.

The effect can be seen on the Overview view for charts that display a
line while they are being loaded, as the CPU Usage chart, when the trace
is large enough to cause multiple  calls of the `fetchTree` endpoint to
the backend. After the data is loaded - for instance after the overview
chart is completely rendered - the data is cached and all similar charts
open quickly.

Currently theia-trace-extension renders one chart at a time. In large
traces, while the first chart is being rendered, if the user opens other
charts they will remain empty until the first one is finished. Work is
being planned to study the possibility of rendering multiple charts in
parallel.

Fixes #816

Signed-off-by: Rodrigo Pinto <rodrigo.pinto@calian.ca>