Add configurable default view for the trace overview
Currently, the trace overview default view is hard-coded in the trace
viewer. This commit allows users/developers to set an XY view as the
default view for the trace overview (using its name).

Users can set the default view for the trace overview using the
Preference UI (File > Preferences > Open Settings or Ctrl + ,).

Developers can also set the default view for a specific Theia build by
adding the "trace Viewer.trace Overview.defaultView" property to the
"preferences" object in the package.json file of the build and indicate
the name of the default view.

Fixes #840.

Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>