Add configurable default view for the trace overview
Currently, the trace overview default view is hard-coded in the trace
viewer. With this commit, the default view for the trace extension can
be set as a Preference setting in Theia. By default, the default view is
Histogram, but that can be changed for each Theia build using the
package.json file.

Fixes #840.

Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>