Toggle the filter tree for the trace overview
This commit creates a hamburger menu for the trace overview
that will allow users to toggle the filter tree and select
a different output view. The hamburger menu replaces the existing
cog icon that is used for output source selection. A new
state property is added to the AbstractTreeOutputComponent
that allows showing and hiding the filter tree. The overview
then toggles this state property.

Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>