Fix overview opened in all trace tabs when using the toolbar button
Currently, the trace overview open signal is broadcasted to all opened
traces which causes the overview to open in all trace tabs. This commit
adds a traceId parameter to the function that opens the trace overview,
so that the trace viewer can check and only open the overview for the
current active trace tab.

Fixes #856.

Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>