Add new command open-trace-with-path to open a trace with the given path
This new command takes a path to a trace directory and opens the
trace in the trace viewer.

The use case is we have an extension that collects traces from
target hardware, when that trace collection is complete we want
to open it in the UI. Adding this new command allows us to open
the trace in the viewer by passing the path to the new trace.

The alternative we considered was, in our extension, connecting
to the trace server and issuing a tspClient.openTrace and then
tspClient.createExperiment. While this does load the trace on
the server, we still need a way for the UI to refresh. Rather
the providing a new command to refresh the UI we propose this
new command as it also saves us from duplicating all the work
to interact with the tspClient.