Website for Theia Trace Extension documentation
Adds a website in the "doc" folder to hold the documentation for Theia
Trace Extension.

Website built with [Svelte Kit](https://kit.svelte.dev).

This PR replaces #776, which accomplishes the same goal but uses
GatsbyJS. The proposal to replace Gatsby with Svelte aims at reducing
the code base and adopting more modern solutions for the problem of
rendering static documentation sites.

Signed-off-by: Rodrigo Pinto <rodrigo.pinto@calian.ca>