Reduce timeline-chart cache by reducing worldRangeFactor to 0.25
With this change only 0.25 times (instead of 1.0) of the view range is
cached on the before and after of the window range. This reduces the
amount of data to be fetched and managed down. This will improve the
performance of the timeline-chart when, for example, zooming.

Horizontal panning inside the world range is still available due to the
cached data.

Fixes #912

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>