JS script to copy ADRs automatically
The script copies ADRs from the `doc/adr` folder into the web-doc
project, so that HTML pages can be generated for them.

The script runs automatically when the user starts the project with
`yarn dev`.

Before copying the files the script compares folders and only copies
folders and files that are missing, or that were modified.

The script runs synchronously, so that the project only starts after the
ADRs have been copied.

Signed-off-by: Rodrigo Pinto <rodrigo.pinto@calian.ca>