Don't use Enzyme in tooltip-component.test.tsx
Use react-testing-library instead. Keep skipping tests due to issues
with the TooltipComponent in tests.

react-tooltip v4.2.14 works in the application but causes an exception
in tests (https://github.com/ReactTooltip/react-tooltip/issues/681),
which is fixed in v4.2.17. However, version v4.2.17 breaks the tooltip
when running in the application, and hence we can't upgrade.

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>