releng: Define publishing latest semantic versions
Add a publish-latest GitHub workflow similar to the existing
publish-next. Make the former trigger on merging a commit that bumps the
package versions depending on the chosen yarn script command. Hence,
make the latter trigger only on the other kinds of commits.

Align publish-next (with publish-latest) also so its previous trivial
nits don't show as differences, upon comparing the two files for review
purposes or so.

Beside the usual goal of publishing onto the npm Registry, use the
github-tag-action in publish-latest, to push the corresponding tag after
that publishing. Make the hereby generated commit (message)
automatically convey how to name that bumped tag. For that purpose,
github-tag-action uses a message token such as #major, #minor or #patch.

Here, publish-latest expects GITHUB_TOKEN to be set, similarly to the
already used NPM_AUTH_TOKEN.

Add the related yarn scripts and document how to use them (commands) in
the CONTRIBUTING markdown. Document also how the aforementioned workflow
is expected to automate the process that includes these commands. And,
mention how to work-around any potential workflow issue manually; if
any until that can be fixed. This adds the benefit of documenting how
that automation proceeds, step-wise.

Fix .gitignore's examples-related lines so they work also upon using the
hereby added yarn commands.

Once confirmed, this approach can be ported to [1] and [2] in turn. With
this approach, introducing a Jenkins job [3] wouldn't be necessary.

[1] https://github.com/eclipse-cdt-cloud/timeline-chart
[2] https://github.com/eclipse-cdt-cloud/tsp-typescript-client
[3] https://ci.eclipse.org/cdt-cloud/

Signed-off-by: Marco Miller <marco.miller@ericsson.com>