Resolve warning errors with the toggle columns menu
This commit makes checkboxes of the toggle columns menu readonly to
resolve the error warning in the developer console. Since the on click
handler is registered to the list item, there is no need to register an on
change handler to the checkbox.

This commit also adds a key to each list item in the title bar menu
drop down since it also triggers an error in the console.

Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>