[lerna] upgrade to recent version, refactoring of package.json "scripts" entries
Updated lerna from v4.x to 7.x, getting rid of several big security vulnerabilities,
that were present in the older version. Also updated ther repo's root README to do
small adjustments, related to the changes. I used the README to test too, replicating
the contained commands.

More details:

root package.json and electron/browser package.json:

Cleaned-up the "scripts" section.

Some uses of lerna in there were not working any more, so they were replaced with a
non-lerna alternative. Since I was already in there, I took the opportunity to
refactor the "scripts" section. I attempted to keep the "public API" the same - i.e.
so the the commands that are documented in the README and used elsewhere in the repo
should still work the same.

For instance:

- Added convenience shortcuts to call scripts in browser and
electron sub-packages, called respectively "electron" and "browser",
rather than having to use the long form every time (yarn --cwd <folder> <command>)
- Moved to the specific application, some scripts that ought to be defined there,
leaving the root entries as shortcuts, to call the specific implemetation(s).
- Introduced the "concurrently" dev-dependency - it helps start two or more processes
from one "script" entry, format their output so it's readable and easy to distinguish
what came out of which process. It also can watch all started processes and make sure
that if one exits, the others are killed too (using the "--kill-others" option). This
is used in the "scripts" entries that start both one of the viewer application (browser
or electron) and the trace server.
- To avoid potential issues, modified the "workspaces" section to specifically name the
browser and electron applications separately rather than use a glob (*): the reason is
that the "plugins" folder is created in the same "examples" folder, and could be
considered a yarn workspace entry on its own, which we do not want.

lerna.json:

updated, according to the new version used.

electron-builder config:

Corrected relative path to the plugins folder, so that it will be included in packages,
such as AppImage and deb on Linux. Also updated currently used Electron version
(23.6.0).

Fixes #1013

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>